+++ hari_client/upload/hari_uploader.py
@@ -12,6 +12,9 @@ from hari_client.utils import logger
 
 log = logger.setup_logger(__name__)
 
+# the maximum attributes number for the whole dataset/upload
+MAX_ATTR_COUNT = 1000
+
 
 class HARIAttribute(models.BulkAttributeCreate):
     # overwrites the annotatable_id and _type fields to not be required,
@@ -81,6 +84,10 @@ class HARIMediaObjectUploadError(Exception):
 
 
 class HARIMediaObjectUnknownObjectCategorySubsetNameError(Exception):
+    pass
+
+
+class HARIAttributesUploadError(Exception):
     pass
 
 
@@ -164,6 +171,7 @@ class HARIUploader:
         self._attribute_cnt: int = 0
         # TODO: this should be a dict[str, uuid.UUID] as soon as the api models are updated
         self._object_category_subsets: dict[str, str] = {}
+        self._unique_attributes: set[uuid.UUID] = set()
 
     # TODO: add_media shouldn't do validation logic, because that expects that a specific order of operation is necessary,
     # specifically that means that media_objects and attributes have to be added to media before the media is added to the uploader.
@@ -193,6 +201,9 @@ class HARIUploader:
 
             self._medias.append(media)
             self._attribute_cnt += len(media.attributes)
+            for attr in media.attributes:
+                self._unique_attributes.add(attr.id)
+
             # check and remember media object back_references
             for media_object in media.media_objects:
                 if media_object.back_reference in self._media_object_back_references:
@@ -206,6 +217,8 @@ class HARIUploader:
                     self._media_object_back_references.add(media_object.back_reference)
                 self._media_object_cnt += 1
                 self._attribute_cnt += len(media_object.attributes)
+                for attr in media_object.attributes:
+                    self._unique_attributes.add(attr.id)
 
     def _add_object_category_subset(self, object_category: str, subset_id: str) -> None:
         self._object_category_subsets[object_category] = subset_id
@@ -371,6 +384,13 @@ class HARIUploader:
                 "before calling HARIUploader::upload()."
             )
             return None
+
+        if len(self._unique_attributes) > MAX_ATTR_COUNT:
+            raise HARIAttributesUploadError(
+                f"You are trying to upload too many attributes for one dataset: {len(self._unique_attributes)}, the limit is {MAX_ATTR_COUNT}. "
+                f"Please, make sure that you don't create unique attribute ids withing all annotatables within the same attribute. "
+                f"See how to create an attribute properly here https://docs.quality-match.com/hari_client/step_by_step/#step-6-create-an-attribute-on-the-image-or-geometry . "
+            )
 
         self._handle_object_categories()